Restructure shoop.core.pricing
Move all code to private submodules.

Refs SHOOP-1942/SHOOP-1955
