+++ shoop/core/pricing/__init__.py
@@ -44,189 +44,24 @@ TODO: caching.
 
 from __future__ import unicode_literals
 
-import abc
-import hashlib
+from shoop.utils import update_module_attributes
 
-import six
-from django.http import HttpRequest
-from django.utils.encoding import force_bytes
-from django.utils.timezone import now
-
-from shoop.apps.provides import load_module
-
-from .price import Price, TaxfulPrice, TaxlessPrice
-from .price_info import PriceInfo
-from .priceful import Priceful
+from ._context import PricingContext, PricingContextable
+from ._module import get_pricing_module, PricingModule
+from ._price import Price, TaxfulPrice, TaxlessPrice
+from ._price_info import PriceInfo
+from ._priceful import Priceful
 
 __all__ = [
     "Price",
     "Priceful",
     "PriceInfo",
     "PricingContext",
+    "PricingContextable",
     "PricingModule",
     "TaxfulPrice",
     "TaxlessPrice",
     "get_pricing_module",
 ]
 
-
-def get_pricing_module():
-    """
-    :rtype: shoop.core.pricing.PricingModule
-    """
-    return load_module("SHOOP_PRICING_MODULE", "pricing_module")()
-
-
-class PricingContextable(six.with_metaclass(abc.ABCMeta)):
-    """
-    Object that is or can be converted to a pricing context.
-
-    Currently there exists two kind of `PricingContextable` objects:
-    `PricingContext`(and its subclasses) and `HttpRequest`.
-
-    .. note::
-
-       Expression ``isinstance(request, PricingContextable)`` will
-       return True for a ``request`` which is `HttpRequest`, because
-       `HttpRequest` is registered as a subclass of this abstract base
-       class.
-
-    This abstract base class is just a helper to allow writing simpler
-    type specifiers, since we want to allow passing `HttpRequest` as a
-    pricing context even though it is not a `PricingContext`.
-    """
-    pass
-PricingContextable.register(HttpRequest)
-
-
-class PricingContext(PricingContextable):
-    """
-    Context for pricing.
-    """
-    REQUIRED_VALUES = ()
-
-    def __init__(self, **kwargs):
-        kwargs.setdefault("time", now())
-        for name, value in kwargs.items():
-            setattr(self, name, value)
-        for name in self.REQUIRED_VALUES:
-            if not hasattr(self, name):
-                raise ValueError("%s is a required value for %s but is not set." % (name, self))
-
-    def get_cache_key_parts(self):
-        return [getattr(self, key) for key in self.REQUIRED_VALUES]
-
-    def get_cache_key(self):
-        parts_text = "\n".join(force_bytes(part) for part in self.get_cache_key_parts())
-        return "%s_%s" % (
-            self.__class__.__name__,
-            hashlib.sha1(parts_text).hexdigest()
-        )
-
-    cache_key = property(get_cache_key)
-
-
-class PricingModule(six.with_metaclass(abc.ABCMeta)):
-    identifier = None
-    name = None
-    pricing_context_class = PricingContext
-
-    def get_context(self, context):
-        """
-        Create pricing context from pricing contextable object.
-
-        :type context: PricingContextable
-        :rtype: PricingContext
-        """
-        if isinstance(context, self.pricing_context_class):
-            return context
-        elif isinstance(context, HttpRequest):
-            return self.get_context_from_request(context)
-        raise TypeError("Not pricing contextable: %r" % (context,))
-
-    def get_context_from_request(self, request):
-        """
-        Create pricing context from HTTP request.
-
-        This base class implementation does not use `request` at all.
-
-        :type request: HttpRequest
-        :rtype: PricingContext
-        """
-        return self.pricing_context_class()
-
-    def get_context_from_data(self, **context_data):
-        """
-        Create pricing context from keyword arguments.
-
-        :rtype: PricingContext
-        """
-        return self.pricing_context_class(**context_data)
-
-    @abc.abstractmethod
-    def get_price_info(self, context, product, quantity=1):
-        """
-        :param product: `Product` object or id of `Product`
-        :type product: shoop.core.models.Product|int
-        :rtype: PriceInfo
-        """
-        pass
-
-    def get_pricing_steps(self, context, product):
-        """
-        Get context-specific list pricing steps for the given product.
-
-        Returns a list of PriceInfos ``[pi0, pi1, pi2, ...]`` where each
-        PriceInfo object is at the border unit price change: unit price
-        for ``0 <= quantity < pi1.quantity1`` is
-        ``pi0.discounted_unit_price``, and unit price for ``pi1.quantity
-        <= quantity < pi2.quantity`` is ``pi1.discounted_unit_price``,
-        and so on.
-
-        If there are "no steps", the return value will be a list of single
-        PriceInfo object with the constant price, i.e. ``[price_info]``.
-
-        :param product: Product or product id
-        :type product: shoop.core.models.Product|int
-        :rtype: list[PriceInfo]
-        """
-        return [self.get_price_info(context, product, quantity=1)]
-
-    def get_price_infos(self, context, products, quantity=1):
-        """
-        Get PriceInfo objects for a bunch of products.
-
-        Returns a dict with product id as key and PriceInfo as value.
-
-        May be faster than doing :func:`get_price_info` for each product
-        separately, since inheriting class may override this.
-
-        :param products: a list of `Product`s or id's
-        :type products:  Iterable[shoop.core.models.Product|int]
-        :rtype: dict[int,PriceInfo]
-        """
-        product_ids = [getattr(x, "pk", x) for x in products]
-        return {
-            product_id: self.get_price_info(context=context, product=product_id, quantity=quantity)
-            for product_id in product_ids
-        }
-
-    def get_pricing_steps_for_products(self, context, products):
-        """
-        Get pricing steps for a bunch of products.
-
-        Returns a dict with product id as key and step data (as list of
-        PriceInfos) as values.
-
-        May be faster than doing :func:`get_pricing_steps` for each
-        product separately, since inheriting class may override this.
-
-        :param products: a list of `Product`s or id's
-        :type products:  Iterable[shoop.core.models.Product|int]
-        :rtype: dict[int,list[PriceInfo]]
-        """
-        product_ids = [getattr(x, "pk", x) for x in products]
-        return {
-            product_id: self.get_pricing_steps(context, product_id=product_id)
-            for product_id in product_ids
-        }
+update_module_attributes(__all__, __name__)
