+++ shoop/core/pricing/_priceful_properties.py
@@ -4,7 +4,7 @@
 #
 # This source code is licensed under the AGPLv3 license found in the
 # LICENSE file in the root directory of this source tree.
-from .price import TaxfulPrice, TaxlessPrice
+from ._price import TaxfulPrice, TaxlessPrice
 
 
 class PriceRate(object):
